# encoding: use lookup table for PeekType