It's faster than the switch statement, which had to do a lot of comparisons. Even making the switch statement not use arbitrary statements doesn't cause the compiler to make a lookup table, so we do it by hand.

```
name         old time/op    new time/op    delta
PeekType-24    4.92ns ± 0%    0.38ns ± 0%  -92.27%  (p=0.000 n=10+10)

name         old alloc/op   new alloc/op   delta
PeekType-24     0.00B          0.00B          ~     (all equal)

name         old allocs/op  new allocs/op  delta
PeekType-24      0.00           0.00          ~     (all equal)
```