@jordanlewis 2018-10-02
TFTR! Enhancing the benchmark doesn't seem important, considering we're now testing lookups into an array, so I'm going to skip it.